Update entity trait
