+++ Entity/EntityTrait/DateTimeControlTrait.php
@@ -4,23 +4,6 @@ namespace Grossum\CoreBundle\Entity\EntityTrait;
 
 trait DateTimeControlTrait
 {
-    /**
-     * Executes automatically before inserting record
-     */
-    public function setCreatedAtValue()
-    {
-        $this->createdAt = new \DateTime();
-
-        return $this;
-    }
-
-    /**
-     * Executes automatically before inserting/updating record
-     */
-    public function setUpdatedAtValue()
-    {
-        $this->updatedAt = new \DateTime();
-
-        return $this;
-    }
+    use CreatedAtControlTrait;
+    use UpdateAtControlTrait;
 }
